requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
google-cloud-firestore
openai
python-dotenv